Update preprocessing.py
It takes the image from the canvas and turns it into a 20*20 binary vector, which the model can understand.
I also added an option to add noise to the input. this helps test how strong the model is with unclear or massy drawings.
The function I made is used in gui.py when the user clicks the classify button.